requirements.txt
python-dotenv
apscheduler
python-telegram-bot==20.7
requests
openai
pandas
matplotlib
numpy